# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthControllerAdapter_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequestModel"}}}},"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Auth"]}},"/users":{"post":{"operationId":"UsersControllerAdapter_create","summary":"","description":"Creacion de usuarios","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersRequestModel"}}}},"responses":{"201":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Users"]}},"/users/{email}":{"post":{"operationId":"UsersControllerAdapter_active","summary":"","description":"Activacion de usuario","parameters":[{"name":"email","required":true,"in":"path","example":"test@gmail.com","schema":{"type":"string"}}],"responses":{"201":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Users"]}},"/products":{"get":{"operationId":"ProductsControllerAdapter_list","summary":"","description":"Listado de productos","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Products"],"security":[{"bearer":[]}]},"post":{"operationId":"ProductsControllerAdapter_create","summary":"","description":"Crea un nuevo producto","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsRequestModel"}}}},"responses":{"201":{"description":""},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Products"],"security":[{"bearer":[]}]}},"/products/pagination":{"get":{"operationId":"ProductsControllerAdapter_listFindQuery","summary":"","description":"Listado de productos","parameters":[{"name":"page","required":true,"in":"query","example":"1","description":"Página","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","description":"Nombre del producto","schema":{"type":"string"}},{"name":"amount","required":false,"in":"query","schema":{"minimum":0.1,"type":"number"}},{"name":"count","required":false,"in":"query","schema":{"minimum":1,"type":"number"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"tags","required":false,"in":"query","schema":{"type":"string"}},{"name":"description","required":false,"in":"query","schema":{"type":"string"}},{"name":"information","required":false,"in":"query","schema":{"type":"string"}},{"name":"assessment","required":false,"in":"query","description":"Valoración del producto [0-5]","schema":{"minimum":0,"maximum":5,"type":"number"}},{"name":"sku","required":false,"in":"query","schema":{"type":"string"}},{"name":"imagen","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Products"],"security":[{"bearer":[]}]}},"/products/count":{"get":{"operationId":"ProductsControllerAdapter_count","summary":"","description":"Listado de productos","parameters":[{"name":"page","required":true,"in":"query","example":"1","description":"Página","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","description":"Nombre del producto","schema":{"type":"string"}},{"name":"amount","required":false,"in":"query","schema":{"minimum":0.1,"type":"number"}},{"name":"count","required":false,"in":"query","schema":{"minimum":1,"type":"number"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"tags","required":false,"in":"query","schema":{"type":"string"}},{"name":"description","required":false,"in":"query","schema":{"type":"string"}},{"name":"information","required":false,"in":"query","schema":{"type":"string"}},{"name":"assessment","required":false,"in":"query","description":"Valoración del producto [0-5]","schema":{"minimum":0,"maximum":5,"type":"number"}},{"name":"sku","required":false,"in":"query","schema":{"type":"string"}},{"name":"imagen","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Products"],"security":[{"bearer":[]}]}},"/products/{sku}":{"put":{"operationId":"ProductsControllerAdapter_update","summary":"","description":"Actualiza la informaión de un producto","parameters":[{"name":"sku","required":true,"in":"path","example":"001","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsRequestUpdateModel"}}}},"responses":{"201":{"description":""},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Products"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductsControllerAdapter_delete","summary":"","description":"Elimina un producto","parameters":[{"name":"sku","required":true,"in":"path","example":"001","schema":{"type":"string"}}],"responses":{"204":{"description":""},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Products"],"security":[{"bearer":[]}]}},"/sales/product":{"post":{"operationId":"SalesControllerAdapter_salesProduct","summary":"","description":"Realiza la venta de un listado de productos","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalesProductModel"}}}},"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Sales"],"security":[{"bearer":[]}]}},"/sales/list":{"get":{"operationId":"SalesControllerAdapter_listSales","summary":"","description":"Listado de articulos vendidos","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Sales"],"security":[{"bearer":[]}]}},"/sales/list-product-zero":{"get":{"operationId":"SalesControllerAdapter_listProductZero","summary":"","description":"Listado de productos con stock en cero","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Sales"],"security":[{"bearer":[]}]}},"/sales/total-earnings":{"get":{"operationId":"SalesControllerAdapter_totalEarning","summary":"","description":"Ganancias totales","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDTO"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/ResponseDTO"}}}]}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"401":{"description":"User Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}},"default":{"description":"Response default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO"}}}}},"tags":["Sales"],"security":[{"bearer":[]}]}}},"info":{"title":"API Rest Test BB Digital","description":"Desarrollador Omar Isalgue Begue","version":"1.0.0","contact":{"name":"BB Digital","url":"http://prueba.com","email":"omar26begue@gmail.com"}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResponseDTO":{"type":"object","properties":{"data":{"type":"object"},"message":{"type":"string"}},"required":["data","message"]},"AuthRequestModel":{"type":"object","properties":{"email":{"type":"string","example":"test@gmail.com"},"password":{"type":"string","example":"123456","minLength":6}},"required":["email","password"]},"UsersRequestModel":{"type":"object","properties":{"email":{"type":"string","example":"test@gmail.com"},"password":{"type":"string","example":"123456","minLength":6},"rol":{"type":"string","enum":["ADMIN_ROL","EDIT_ROL","USER_ROL"],"example":"USER_ROL","default":"USER_ROL"},"name":{"type":"string","example":"John Gonzalez"}},"required":["email","password","rol","name"]},"ProductsRequestModel":{"type":"object","properties":{"name":{"type":"string","example":"Producto de ejemplo","description":"Nombre del producto"},"amount":{"type":"number","example":"12345.12","minimum":0.1},"count":{"type":"number","example":"123","minimum":1},"category":{"type":"string","example":"Prueba"},"tags":{"type":"string","example":"Ventas"},"description":{"type":"string","example":"Ejemplo de descripción producto"},"information":{"type":"string","example":"Información del producto"},"assessment":{"type":"number","example":"4","minimum":0,"maximum":5,"description":"Valoración del producto [0-5]"},"sku":{"type":"string","example":"001"},"imagen":{"type":"string","example":"http://www.example.com/"}},"required":["name","amount","count","category","tags","description","information","assessment","sku","imagen"]},"ProductsRequestUpdateModel":{"type":"object","properties":{"name":{"type":"string","example":"Producto de ejemplo","description":"Nombre del producto"},"amount":{"type":"number","example":"12345.12","minimum":0.1},"count":{"type":"number","example":"123","minimum":1},"category":{"type":"string","example":"Prueba"},"tags":{"type":"string","example":"Ventas"},"description":{"type":"string","example":"Ejemplo de descripción producto"},"information":{"type":"string","example":"Información del producto"},"assessment":{"type":"number","minimum":0,"maximum":5,"description":"Valoración del producto [0-5]"},"imagen":{"type":"string","example":"http://www.example.com/"}},"required":["name","amount","count","category","tags","description","information","assessment","imagen"]},"SalesProductModel":{"type":"object","properties":{"products":{"example":"[\"001\",\"002\"]","description":"Listado de productos","type":"array","items":{"type":"string"}}},"required":["products"]}}}}